Update config.js
fix(js): Prevent crash on shortcut initialization

Handle cases where `pluginConfig.Shortcuts` is not yet defined during the initialization of the script. #26

This prevents a `TypeError` that could occur if the plugin configuration is slow to load from the server by providing an empty array as a fallback. This ensures the rest of the plugin can continue to load and function even if the shortcut configuration is delayed.